Make `make_derivative` and `make_derivative2` public methods
Also moved the methods to the `common` module in the
`pde.grids.operators` package and adjusted their signature such that
they can take any grid.
